Refactor(#62): common 컴포넌트 ui 컴포넌트로 교체
